split functions.h into .h. and .c
This makes including it in several locations possible without linker errors
