+++ CMakeLists.txt
@@ -1,4 +1,4 @@
 idf_component_register(
-  SRCS "src/esp_dmx.c" "src/esp_rdm.c"
+  SRCS "src/esp_dmx.c" "src/esp_rdm.c" "src/private/rdm_encode/functions.c"
   INCLUDE_DIRS "src" 
 )
